+++ wagtail/contrib/routable_page/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'wagtail.contrib.routable_page.apps.WagtailRoutablePageAppConfig'
