+++ wagtail/contrib/routable_page/apps.py
@@ -4,6 +4,6 @@ from django.apps import AppConfig
 
 
 class WagtailRoutablePageAppConfig(AppConfig):
-    name = 'wagtail.contrib.wagtailroutablepage'
+    name = 'wagtail.contrib.routable_page'
     label = 'wagtailroutablepage'
     verbose_name = "Wagtail routablepage"
